Move Optimizers, update to add x_init and remove unwanted args (#55)
* Move Optimizers

* Fix PEP8 Errors

* Prevent stack example from running temporarily

* Fix line doc
