+++ examples/non_cartesian_reconstruction_stack_3d.py
@@ -15,7 +15,7 @@ We use the toy datasets available in pysap, more specifically a 3D Orange.
 
 # Package import
 from mri.operators import Stacked3DNFFT
-from mri.numerics.reconstruct import sparse_rec_fista
+from mri.optimizers import fista
 from mri.reconstruct.utils import generate_operators
 from mri.operators.utils import convert_locations_to_mask, get_stacks_fourier
 from mri.parallel_mri.extract_sensitivity_maps import \
@@ -96,7 +96,7 @@ gradient_op, linear_op, prox_op, cost_op = generate_operators(
 
 # Start the FISTA reconstruction
 max_iter = 10
-x_final, costs, metrics = sparse_rec_fista(
+x_final, costs, metrics = fista(
     gradient_op=gradient_op,
     linear_op=linear_op,
     prox_op=prox_op,
