+++ mri/optimizers/__init__.py
@@ -7,9 +7,5 @@
 # for details.
 ##########################################################################
 
-"""
-This module contains shortcuts to reweight classes.
-"""
-
-# Package import
-from mri.reconstruct.reweight import mReweight
+from .forward_backward import fista, pogm
+from .primal_dual import condatvu
