+++ mri/optimizers/utils/__init__.py
@@ -6,6 +6,3 @@
 # http://www.cecill.info/licences/Licence_CeCILL-B_V1-en.html
 # for details.
 ##########################################################################
-
-""" This module defines shortcuts to common tools.
-"""
